로또 당첨 번호 및 회차 출력 함수 구현
